+++ src/CodeCoverage.php
@@ -28,12 +28,9 @@ use ReflectionClass;
 use SebastianBergmann\CodeCoverage\Driver\Driver;
 use SebastianBergmann\CodeCoverage\Node\Builder;
 use SebastianBergmann\CodeCoverage\Node\Directory;
-use SebastianBergmann\CodeCoverage\StaticAnalysis\CachingCoveredFileAnalyser;
-use SebastianBergmann\CodeCoverage\StaticAnalysis\CachingUncoveredFileAnalyser;
-use SebastianBergmann\CodeCoverage\StaticAnalysis\CoveredFileAnalyser;
-use SebastianBergmann\CodeCoverage\StaticAnalysis\ParsingCoveredFileAnalyser;
-use SebastianBergmann\CodeCoverage\StaticAnalysis\ParsingUncoveredFileAnalyser;
-use SebastianBergmann\CodeCoverage\StaticAnalysis\UncoveredFileAnalyser;
+use SebastianBergmann\CodeCoverage\StaticAnalysis\CachingFileAnalyser;
+use SebastianBergmann\CodeCoverage\StaticAnalysis\FileAnalyser;
+use SebastianBergmann\CodeCoverage\StaticAnalysis\ParsingFileAnalyser;
 use SebastianBergmann\CodeUnitReverseLookup\Wizard;
 
 /**
@@ -68,9 +65,7 @@ final class CodeCoverage
      */
     private array $parentClassesExcludedFromUnintentionallyCoveredCodeCheck = [];
 
-    private ?CoveredFileAnalyser $coveredFileAnalyser = null;
-
-    private ?UncoveredFileAnalyser $uncoveredFileAnalyser = null;
+    private ?FileAnalyser $analyser = null;
 
     private ?string $cacheDirectory = null;
 
@@ -87,7 +82,7 @@ final class CodeCoverage
      */
     public function getReport(): Directory
     {
-        return (new Builder($this->coveredFileAnalyser()))->build($this);
+        return (new Builder($this->analyser()))->build($this);
     }
 
     /**
@@ -404,7 +399,7 @@ final class CodeCoverage
         foreach (array_keys($data->lineCoverage()) as $filename) {
             $data->keepLineCoverageDataOnlyForLines(
                 $filename,
-                $this->uncoveredFileAnalyser()->executableLinesIn($filename)
+                $this->analyser()->executableLinesIn($filename)
             );
         }
     }
@@ -418,7 +413,7 @@ final class CodeCoverage
 
             $data->removeCoverageDataForLines(
                 $filename,
-                $this->coveredFileAnalyser()->ignoredLinesFor($filename)
+                $this->analyser()->ignoredLinesFor($filename)
             );
         }
     }
@@ -438,7 +433,7 @@ final class CodeCoverage
                 $this->append(
                     RawCodeCoverageData::fromUncoveredFile(
                         $uncoveredFile,
-                        $this->uncoveredFileAnalyser()
+                        $this->analyser()
                     ),
                     self::UNCOVERED_FILES
                 );
@@ -552,42 +547,24 @@ final class CodeCoverage
         return array_values($unintentionallyCoveredUnits);
     }
 
-    private function coveredFileAnalyser(): CoveredFileAnalyser
+    private function analyser(): FileAnalyser
     {
-        if ($this->coveredFileAnalyser !== null) {
-            return $this->coveredFileAnalyser;
+        if ($this->analyser !== null) {
+            return $this->analyser;
         }
 
-        $this->coveredFileAnalyser = new ParsingCoveredFileAnalyser(
+        $this->analyser = new ParsingFileAnalyser(
             $this->useAnnotationsForIgnoringCode,
             $this->ignoreDeprecatedCode
         );
 
         if ($this->cachesStaticAnalysis()) {
-            $this->coveredFileAnalyser = new CachingCoveredFileAnalyser(
+            $this->analyser = new CachingFileAnalyser(
                 $this->cacheDirectory,
-                $this->coveredFileAnalyser
+                $this->analyser
             );
         }
 
-        return $this->coveredFileAnalyser;
-    }
-
-    private function uncoveredFileAnalyser(): UncoveredFileAnalyser
-    {
-        if ($this->uncoveredFileAnalyser !== null) {
-            return $this->uncoveredFileAnalyser;
-        }
-
-        $this->uncoveredFileAnalyser = new ParsingUncoveredFileAnalyser;
-
-        if ($this->cachesStaticAnalysis()) {
-            $this->uncoveredFileAnalyser = new CachingUncoveredFileAnalyser(
-                $this->cacheDirectory,
-                $this->uncoveredFileAnalyser
-            );
-        }
-
-        return $this->uncoveredFileAnalyser;
+        return $this->analyser;
     }
 }