+++ src/StaticAnalysis/ParsingFileAnalyser.php
@@ -30,7 +30,7 @@ use SebastianBergmann\LinesOfCode\LinesOfCode;
 /**
  * @internal This class is not covered by the backward compatibility promise for phpunit/php-code-coverage
  */
-final class ParsingCoveredFileAnalyser implements CoveredFileAnalyser
+final class ParsingFileAnalyser implements FileAnalyser
 {
     private array $classes = [];
 
@@ -44,6 +44,8 @@ final class ParsingCoveredFileAnalyser implements CoveredFileAnalyser
     private array $linesOfCode = [];
 
     private array $ignoredLines = [];
+
+    private array $executableLines = [];
 
     private bool $useAnnotationsForIgnoringCode;
 
@@ -83,6 +85,13 @@ final class ParsingCoveredFileAnalyser implements CoveredFileAnalyser
         return $this->linesOfCode[$filename];
     }
 
+    public function executableLinesIn(string $filename): array
+    {
+        $this->analyse($filename);
+
+        return $this->executableLines[$filename];
+    }
+
     public function ignoredLinesFor(string $filename): array
     {
         $this->analyse($filename);
@@ -116,16 +125,18 @@ final class ParsingCoveredFileAnalyser implements CoveredFileAnalyser
 
             assert($nodes !== null);
 
-            $traverser                  = new NodeTraverser;
-            $codeUnitFindingVisitor     = new CodeUnitFindingVisitor;
-            $lineCountingVisitor        = new LineCountingVisitor($linesOfCode);
-            $ignoredLinesFindingVisitor = new IgnoredLinesFindingVisitor($this->useAnnotationsForIgnoringCode, $this->ignoreDeprecatedCode);
+            $traverser                     = new NodeTraverser;
+            $codeUnitFindingVisitor        = new CodeUnitFindingVisitor;
+            $lineCountingVisitor           = new LineCountingVisitor($linesOfCode);
+            $ignoredLinesFindingVisitor    = new IgnoredLinesFindingVisitor($this->useAnnotationsForIgnoringCode, $this->ignoreDeprecatedCode);
+            $executableLinesFindingVisitor = new ExecutableLinesFindingVisitor;
 
             $traverser->addVisitor(new NameResolver);
             $traverser->addVisitor(new ParentConnectingVisitor);
             $traverser->addVisitor($codeUnitFindingVisitor);
             $traverser->addVisitor($lineCountingVisitor);
             $traverser->addVisitor($ignoredLinesFindingVisitor);
+            $traverser->addVisitor($executableLinesFindingVisitor);
 
             /* @noinspection UnusedFunctionResultInspection */
             $traverser->traverse($nodes);
@@ -143,11 +154,12 @@ final class ParsingCoveredFileAnalyser implements CoveredFileAnalyser
         }
         // @codeCoverageIgnoreEnd
 
-        $this->classes[$filename]      = $codeUnitFindingVisitor->classes();
-        $this->traits[$filename]       = $codeUnitFindingVisitor->traits();
-        $this->functions[$filename]    = $codeUnitFindingVisitor->functions();
-        $this->linesOfCode[$filename]  = $lineCountingVisitor->result();
-        $this->ignoredLines[$filename] = [];
+        $this->classes[$filename]         = $codeUnitFindingVisitor->classes();
+        $this->traits[$filename]          = $codeUnitFindingVisitor->traits();
+        $this->functions[$filename]       = $codeUnitFindingVisitor->functions();
+        $this->linesOfCode[$filename]     = $lineCountingVisitor->result();
+        $this->executableLines[$filename] = $executableLinesFindingVisitor->executableLines();
+        $this->ignoredLines[$filename]    = [];
 
         $this->findLinesIgnoredByLineBasedAnnotations($filename, $source, $this->useAnnotationsForIgnoringCode);