+++ tests/tests/RawCodeCoverageDataTest.php
@@ -10,7 +10,7 @@
 namespace SebastianBergmann\CodeCoverage;
 
 use function array_keys;
-use SebastianBergmann\CodeCoverage\StaticAnalysis\ParsingUncoveredFileAnalyser;
+use SebastianBergmann\CodeCoverage\StaticAnalysis\ParsingFileAnalyser;
 
 final class RawCodeCoverageDataTest extends TestCase
 {
@@ -229,9 +229,10 @@ final class RawCodeCoverageDataTest extends TestCase
                 12,
                 14,
                 15,
+                16,
                 18,
             ],
-            array_keys(RawCodeCoverageData::fromUncoveredFile($file, new ParsingUncoveredFileAnalyser)->lineCoverage()[$file])
+            array_keys(RawCodeCoverageData::fromUncoveredFile($file, new ParsingFileAnalyser(true, true))->lineCoverage()[$file])
         );
     }
 
@@ -243,7 +244,7 @@ final class RawCodeCoverageDataTest extends TestCase
             [
                 12,
             ],
-            array_keys(RawCodeCoverageData::fromUncoveredFile($file, new ParsingUncoveredFileAnalyser)->lineCoverage()[$file])
+            array_keys(RawCodeCoverageData::fromUncoveredFile($file, new ParsingFileAnalyser(true, true))->lineCoverage()[$file])
         );
     }
 
@@ -256,9 +257,10 @@ final class RawCodeCoverageDataTest extends TestCase
                 7,
                 9,
                 10,
+                11,
                 13,
             ],
-            array_keys(RawCodeCoverageData::fromUncoveredFile($file, new ParsingUncoveredFileAnalyser)->lineCoverage()[$file])
+            array_keys(RawCodeCoverageData::fromUncoveredFile($file, new ParsingFileAnalyser(true, true))->lineCoverage()[$file])
         );
     }
 
@@ -277,7 +279,7 @@ final class RawCodeCoverageDataTest extends TestCase
                 33,
                 35,
             ],
-            array_keys(RawCodeCoverageData::fromUncoveredFile($file, new ParsingUncoveredFileAnalyser)->lineCoverage()[$file])
+            array_keys(RawCodeCoverageData::fromUncoveredFile($file, new ParsingFileAnalyser(true, true))->lineCoverage()[$file])
         );
     }
 
@@ -306,8 +308,18 @@ final class RawCodeCoverageDataTest extends TestCase
                 54,
                 60,
                 71,
+                83,
+                85,
+                87,
+                89,
+                91,
+                93,
+                95,
+                97,
+                99,
+                101,
             ],
-            array_keys(RawCodeCoverageData::fromUncoveredFile($file, new ParsingUncoveredFileAnalyser)->lineCoverage()[$file])
+            array_keys(RawCodeCoverageData::fromUncoveredFile($file, new ParsingFileAnalyser(true, true))->lineCoverage()[$file])
         );
     }